moves service into main
