+++ cmd/api/main.go
@@ -3,11 +3,14 @@ package main
 import (
 	"context"
 	"log"
+	"sync"
+	"time"
 
 	"github.com/ninnemana/drudge"
 	"github.com/ninnemana/drudge/telemetry"
-
-	"github.com/ninnemana/rpc-demo/pkg/service"
+	"github.com/ninnemana/rpc-demo/pkg/vinyltappb"
+	"github.com/pkg/errors"
+	"google.golang.org/grpc"
 )
 
 const (
@@ -23,7 +26,7 @@ var (
 			Network: "tcp",
 			Addr:    rpcAddr,
 		},
-		OnRegister: service.Register,
+		OnRegister:    Register,
 		TraceExporter: telemetry.Jaeger,
 		TraceConfig: telemetry.JaegerConfig{
 			ServiceName: "rpc-demo",
@@ -35,4 +38,43 @@ func main() {
 	if err := drudge.Run(context.Background(), options); err != nil {
 		log.Fatalf("Fell out of serving application: %+v", err)
 	}
+}
+
+type Service struct {
+	albums map[int32]*vinyltappb.Album
+	sync.RWMutex
+}
+
+func Register(server *grpc.Server) error {
+	vinyltappb.RegisterTapServer(server, &Service{})
+	return nil
+}
+
+func (s *Service) GetAlbum(a *vinyltappb.Album, srv vinyltappb.Tap_GetAlbumServer) error {
+	for k := range s.albums {
+		if k != a.GetId() {
+			continue
+		}
+
+		if err := srv.Send(s.albums[k]); err != nil {
+			return errors.WithMessage(err, "failed to send album over TCP connection")
+		}
+	}
+
+	return nil
+}
+
+func (s *Service) Set(ctx context.Context, a *vinyltappb.Album) (*vinyltappb.Album, error) {
+	if _, ok := s.albums[a.GetId()]; ok {
+		return nil, errors.Errorf("provided album '%d' exists", a.GetId())
+	}
+	s.Lock()
+	s.albums[a.GetId()] = a
+	s.Unlock()
+
+	return s.albums[a.GetId()], nil
+}
+
+func sinceInMilliseconds(startTime time.Time) float64 {
+	return float64(time.Since(startTime).Nanoseconds()) / 1e6
 }
